Make loading ROM to a bad address an error.
Currently if you try to load a ROM into a non-existent region
it is silently ignored. This changes such a case in to a hard
error instead. This ensures that if an ELF file with bad
addresses is loaded, the user knows about it, rather than
having to guess what the problem is.