Provide --disable-console option
It is not always desirable to compile Qemu with a graphical
console. --disable-console disables the Qemu console.

This has a relatively large impact on machines and devices that
assume the existence of a graphical console. The option is only
useful if you are compiling a specific sub-set of machines that
don't use the graphical console.

Console currently includes input as well as output. This could
potentially be refactored to remove this dependency.