Load SP/PC directly from physical memory on ARMv7M.
Previously the SP/PC was loaded indirectly from any registered
ROMs. This is not ideal as the flash may be modified at runtime
by either the application or a debugger, and on a reset the
updated SP/PC should be used.

To enable this to work correctly the CPU must be initialised after
memory has been initialised in the system (otherwise the CPU
is unable to read from memory, causing a segfault.)